refactor : vo로 명명한것을 DTO로 변경
